requirements.txt
apache-airflow==2.5.0  # Version d'Airflow
pandas==1.3.3          # Pour gérer les DataFrames
praw==7.5.0            # API Python pour Reddit
python-dotenv==0.19.2  # Pour gérer les variables d'environnement
boto3==1.18.65         # Pour interagir avec AWS S3
